#95 refactor to use Magento_MediaStorage
